+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v2.3.0
+
+### Added
+
+- "Auto fines" reference ([source][gibdd_fines])
+- Static method `::getAutoFines()`
+
 ## v2.2.0
 
 ### Changed
@@ -8,3 +15,5 @@
 - Required minimal PHPUnit version now `5.7.10`
 - Removed unimportant PHPDoc blocks
 - Code a little bit refactored
+
+[gibdd_fines]:https://xn--90adear.xn--p1ai/mens/fines?page=1